+++ app/app.global.css
@@ -1,0 +1,97 @@
+@import '~antd/dist/antd.css';
+
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+}
+
+div#root {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+.create-project {
+  text-align: center;
+}
+.create-project-header {
+  background-color: white;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+}
+
+Button {
+  padding: 15px;
+  display: inline-block;
+  font-size: 16px;
+  margin: 5px 10px 5px 10px;
+  cursor: pointer;
+}
+
+div.container {
+  height: 100%;
+  display: flex;
+  margin: 0;
+  padding: 0;
+  max-width: inherit;
+}
+
+div.left-menu {
+  height: 100%;
+  width: 100px;
+  background-color: rgb(71, 83, 138);
+}
+button.left-img-buttons {
+  width: 100%;
+  left: 0;
+  right: 0;
+  margin-left: auto;
+  margin-right: auto;
+  background-color: transparent;
+}
+button.keyboard {
+  width: 100%;
+  left: 0;
+  right: 0;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+div.working-area {
+  display: flex;
+  height: 100%;
+  background-color: aliceblue;
+  flex: 1;
+}
+div.main-view {
+  height: 100%;
+  background-color: aliceblue;
+  flex: 1;
+}
+div.grid-view {
+  height: 100%;
+  width: 100%;
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  background-color: aliceblue;
+  flex: 1;
+  overflow-x: hidden;
+  overflow-y: scroll;
+}
+
+div.label-menu {
+  height: 100%;
+  flex-basis: 200px;
+  background-color: rgb(51, 88, 255);
+}
