+++ app/components/labels.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function LabelMenu() {
+  return (
+    <div className="label-menu">
+      <div className="label-infos"></div>
+      <button type="button" id="add-label" className="add-label">
+        Add Label
+      </button>
+    </div>
+  );
+}
+
+export default LabelMenu;
+
