+++ app/main.dev.js
@@ -1,0 +1,90 @@
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
+import path from 'path';
+import { app, BrowserWindow, Menu, ipcMain } from 'electron';
+import { autoUpdater } from 'electron-updater';
+import log from 'electron-log';
+
+const menu = require('./main/menu.js');
+const selectDir = require('./main/select-dirs.js');
+const setProjectManager = require('./main/set-project-manager.js');
+
+export default class AppUpdater {
+  constructor() {
+    log.transports.file.level = 'info';
+    autoUpdater.logger = log;
+    autoUpdater.checkForUpdatesAndNotify();
+  }
+}
+
+let mainWindow = null;
+
+if (process.env.NODE_ENV === 'production') {
+  const sourceMapSupport = require('source-map-support');
+  sourceMapSupport.install();
+}
+if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === 'true') {
+  require('electron-debug')();
+}
+const installExtensions = async () => {
+  const installer = require('electron-devtools-installer');
+  const forceDownload = !!process.env.UPGRADE_EXTENSIONS;
+  const extensions = ['REACT_DEVELOPER_TOOLS', 'REDUX_DEVTOOLS'];
+  return Promise.all(
+    extensions.map((name) => installer.default(installer[name], forceDownload))
+  ).catch(console.log);
+};
+const createWindow = async () => {
+  if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === 'true') {
+    await installExtensions();
+  }
+  mainWindow = new BrowserWindow({
+    show: false,
+    width: 1024,
+    height: 728,
+    webPreferences:
+      (process.env.NODE_ENV === 'development' || process.env.E2E_BUILD === 'true') &&
+      process.env.ERB_SECURE !== 'true'
+        ? {
+            nodeIntegration: true,
+          }
+        : {
+            preload: path.join(__dirname, 'dist/renderer.prod.js'),
+          },
+  });
+  mainWindow.loadURL(`file://${__dirname}/index.html`);
+  mainWindow.webContents.on('did-finish-load', () => {
+    if (!mainWindow) {
+      throw new Error('"mainWindow" is not defined');
+    }
+    if (process.env.START_MINIMIZED) {
+      mainWindow.minimize();
+    } else {
+      mainWindow.show();
+      mainWindow.focus();
+    }
+  });
+  mainWindow.on('closed', () => {
+    mainWindow = null;
+  });
+  Menu.setApplicationMenu(menu);
+};
+
+ipcMain.on('selectDir', selectDir);
+ipcMain.on('setProjectManager', setProjectManager);
+
+app.on('window-all-closed', () => {
+  if (process.platform !== 'darwin') {
+    app.quit();
+  }
+});
+
+if (process.env.E2E_BUILD === 'true') {
+  app.whenReady().then(createWindow);
+} else {
+  app.on('ready', createWindow);
+}
+
+app.on('activate', () => {
+  if (mainWindow === null) createWindow();
+});
