+++ app/utils/exts.js
@@ -1,0 +1,3 @@
+const imgExtensions = ['.png', '.jpg', '.jpeg', '.PNG', '.JPG', '.JPEG'];
+
+module.exports = { imgExtensions };
