--- craco.config.js
@@ -1,17 +1,0 @@
-const CracoLessPlugin = require('craco-less');
-
-module.exports = {
-  plugins: [
-    {
-      plugin: CracoLessPlugin,
-      options: {
-        lessLoaderOptions: {
-          lessOptions: {
-            modifyVars: { '@primary-color': '#80C7F2' },
-            javascriptEnabled: true,
-          },
-        },
-      },
-    },
-  ],
-};
