--- src/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './styles/index.css';
-import CreateProject from './components/create-project';
-
-ReactDOM.render(<CreateProject />, document.getElementById('root'));
