--- src/utils/exts.js
@@ -1,3 +1,0 @@
-const imgExtensions = ['.png', '.jpg', '.jpeg', '.PNG', '.JPG', '.JPEG'];
-
-module.exports = { imgExtensions };
